+++ pkg/models/mines.go
@@ -28,9 +28,9 @@ import (
 )
 
 type MineModel struct {
-	ctx  context.Context
-	s    *zap.SugaredLogger
-	conn *sqlx.Conn
+	ctx context.Context
+	s   *zap.SugaredLogger
+	db  *sqlx.DB
 }
 
 type Mine struct {
@@ -40,8 +40,8 @@ type Mine struct {
 }
 
 // NewMineModel creates a new instance of the 'Mine' Model.
-func NewMineModel(ctx context.Context, s *zap.SugaredLogger, conn *sqlx.Conn) *MineModel {
-	return &MineModel{ctx: ctx, s: s, conn: conn}
+func NewMineModel(ctx context.Context, s *zap.SugaredLogger, db *sqlx.DB) *MineModel {
+	return &MineModel{ctx: ctx, s: s, db: db}
 }
 
 // GetMineIdsByPurlType retrieves a list of the Purl Type IDs associated with the given Purl Type (string).
@@ -51,7 +51,7 @@ func (m *MineModel) GetMineIdsByPurlType(purlType string) ([]int32, error) {
 		return nil, errors.New("please specify a Purl Type to query")
 	}
 	var mines []Mine
-	err := m.conn.SelectContext(m.ctx, &mines,
+	err := m.db.SelectContext(m.ctx, &mines,
 		"SELECT id,mine_name,purl_type FROM mines WHERE purl_type = $1", purlType,
 	)
 	if err != nil {