+++ pkg/models/mines_test.go
@@ -42,7 +42,7 @@ func TestMines(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed to load SQL test data: %v", err)
 	}
-	mine := NewMineModel(ctx, s, conn)
+	mine := NewMineModel(ctx, s, db)
 	var purlType = "maven"
 	mineIds, err := mine.GetMineIdsByPurlType(purlType)
 	if err != nil {
@@ -93,7 +93,7 @@ func TestMinesBadSql(t *testing.T) {
 	defer CloseDB(db)
 	conn := sqliteConn(t, ctx, db) // Get a connection from the pool
 	defer CloseConn(conn)
-	mine := NewMineModel(ctx, s, conn)
+	mine := NewMineModel(ctx, s, db)
 	purlType := "NONEXISTENT"
 	mineIds, err := mine.GetMineIdsByPurlType(purlType)
 	if err != nil {